Tileable background image
The previous image was stretched to fill 100% of the window. This was
not optimal, as this was poorly supported by some browsers, it could
show a stretch pattern and the behaviour on resize was faulty. Now the
image has been generated for being tileable, and is rendered on native
resolution.